refactor(admin): 최종 URL 조립 유틸을 홍보·활동 사진이 함께 쓰게 한다
#2026이 활동 사진에 넣은 buildFinalUrls와 홍보 폼의 orderedUrls가 같은
불변식(결과는 화면 순서의 부분 수열)을 각자 구현하고 있었다.

reorderItems가 이미 있는 components/ImageSortGrid로 옮겨 한 벌만 남긴다.
