+++ frontend/src/pages/AdminPage/tabs/PhotoEditTab/photoEditUtils.ts
@@ -26,19 +26,3 @@ export const hasPendingChanges = (
 
 export const extractLocalItems = (feedItems: ImageItem[]): LocalItem[] =>
   feedItems.filter((item): item is LocalItem => item.type === 'local');
-
-// 화면 순서(feedItems)를 그대로 보존한 최종 URL 배열을 만든다.
-// 업로드되지 않아 URL이 없는 local 아이템은 제외된다.
-export const buildFinalUrls = (
-  feedItems: ImageItem[],
-  urlByFile: Map<File, string>,
-): string[] =>
-  feedItems.reduce<string[]>((acc, item) => {
-    if (item.type === 'uploaded') {
-      acc.push(item.url);
-      return acc;
-    }
-    const url = urlByFile.get(item.file);
-    if (url) acc.push(url);
-    return acc;
-  }, []);
